+++ app/src/main/java/com/suonk/oc_project9/utils/sort/Sorting.kt
@@ -1,6 +1,7 @@
 package com.suonk.oc_project9.utils.sort
 
 import com.suonk.oc_project9.model.database.data.entities.real_estate.RealEstateEntityWithPhotos
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 enum class Sorting(val comparator: Comparator<RealEstateEntityWithPhotos>) {
@@ -12,7 +13,7 @@ enum class Sorting(val comparator: Comparator<RealEstateEntityWithPhotos>) {
     CHRONOLOGICAL_ASC(Comparator.comparing { it.realEstateEntity.id }), CHRONOLOGICAL_DESC(
         Comparator.comparing { it.realEstateEntity.id }),
 
-    PRICE_ASC(Comparator.comparing { it.realEstateEntity.price }), PRICE_DESC(Comparator.comparing<RealEstateEntityWithPhotos?, Double?> { it.realEstateEntity.price }
+    PRICE_ASC(Comparator.comparing { it.realEstateEntity.price }), PRICE_DESC(Comparator.comparing<RealEstateEntityWithPhotos?, BigDecimal?> { it.realEstateEntity.price }
         .reversed()),
 
     LIVING_SPACE_ASC(Comparator.comparing { it.realEstateEntity.livingSpace }), LIVING_SPACE_DESC(
